requirements.txt
fastapi
openai
requests
uvicorn
python-dotenv